requirements.txt
openai-chatkit>=1.0.2
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
aiohttp>=3.9.0
asyncpg>=0.29.0
redis>=5.0.1
python-dotenv>=1.0.0
pydantic>=2.10,<3.0
beautifulsoup4>=4.12.2
httpx>=0.25.2
psycopg2-binary>=2.9.9
